Restructured app to use context at top level
