+++ client/src/components/App.jsx
@@ -1,7 +1,15 @@
 import React from 'react';
 import './styles/App.css';
+import Login from './pages/Login';
 import { Route, Redirect, Link, BrowserRouter as Router } from 'react-router-dom';
-import {userContext} from './userContext';
+import {userContext} from '../userContext.js';
+import ReactDOM from 'react-dom';
+import '../index.css';
+import Auth from './Auth';
+import Main from './pages/Main'
+
+
+
 
 class App extends React.Component {
   constructor(props) {
@@ -15,14 +23,20 @@ class App extends React.Component {
     // get and set currently logged in user to state
   }
 
+  
   render() {
     return (
-      // Pass user state as value to context.Provider so it can be consumed by context.Consumer
-      <userContext.Provider value={this.state.user}>
-        <Main/>
+      <userContext.Provider>
+        <div className="App"></div>
       </userContext.Provider>
+     
     );
   }
 }
 
+
 export default App;
+
+
+
+
