+++ client/src/userContext.js
@@ -1,3 +1,5 @@
+import React from 'react';
+
 const userContext = React.createContext({user: {}}); // Create a context object
 
 export {
